include many changes
Update “ambiguous question” to red, “unambigous question to “green” color
Having a text under the image showing what labels are available.  Also when no question is selected, the UI should be the same as when a question as selected, so the user can see what segmentations are selectable.
Need to have a “previous button” so worker can navigate between the ambiguous question and unambiguous one
Calculate time usage
Add “clear all” button